Move getLiveTest out of the timeline
getLiveTest is test-runtime plumbing, not timeline code. The frame
counter that lands next needs it, and pulling it from the timeline
would create a cycle: the timeline depends on tree snapshots, which
will depend on the counter.
